Change type for "depth,d" parameter from uint8_t to int

+++ CmdLineOptions.cpp
@@ -9,7 +9,7 @@ bool CmdLineOptions::FillFromArgs(int argc, char** argv)
 {
     try
     {
-        auto checkDepthValue = [](const uint8_t& value)
+        auto checkDepthValue = [](const int& value)
             {
                 if (value != 0 && value != 1) throw po::validation_error(po::validation_error::invalid_option_value, "depth");
             };
@@ -33,7 +33,7 @@ bool CmdLineOptions::FillFromArgs(int argc, char** argv)
             ("help,h", "help message")
             ("include-dirs,i", po::value<string_s>()->default_value(string_s{ "." }, ".")->multitoken()->composing(), "Including to search directories, can be multiple. Default is current folder")
             ("exclude-dirs,e", po::value<string_s>()->multitoken()->composing(), "Excluded to search directories, can be multiple")
-            ("depth,d", po::value<uint8_t>()->default_value(0)->notifier(checkDepthValue), "Scan depth, 1 - all directories, 0 - current folder only. Default value is 0")
+            ("depth,d", po::value<int>()->default_value(0)->notifier(checkDepthValue), "Scan depth, 1 - all directories, 0 - current folder only. Default value is 0")
             ("min-size,m", po::value<unsigned>()->default_value(1)->notifier(checkSizeValue), "Minimum size of the processed file in bytes. Default value is 1")
             ("file-masks,f", po::value<string_s>()->multitoken()->composing(), "Masks for the files participating in comparison")
             ("block-size,b", po::value<int>()->default_value(256)->notifier(checkSizeValue), "The size in bytes of the block to read the files with. Default value is 256")
@@ -54,7 +54,7 @@ bool CmdLineOptions::FillFromArgs(int argc, char** argv)
         if (vm.count("file-masks"))
             _fileMasks = vm["file-masks"].as<string_s>();
 
-        _scanLevel = ScanLevel(vm["depth"].as<uint8_t>());
+        _scanLevel = ScanLevel(vm["depth"].as<int>());
         _minFileSize = vm["min-size"].as<unsigned>();
         _blockSize = vm["block-size"].as<int>();
